+++ ARORI/src/main/java/com/kh/arori/interceptor/MemberInterceptor.java
@@ -9,7 +9,7 @@ import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 
 import com.kh.arori.entity.MemberDto;
 
-public class AuthInterceptor extends HandlerInterceptorAdapter {
+public class MemberInterceptor extends HandlerInterceptorAdapter {
 
 	@Override
 	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
@@ -23,24 +23,24 @@ public class AuthInterceptor extends HandlerInterceptorAdapter {
 		// 2. 형 변환
 		HandlerMethod handlerMethod = (HandlerMethod) handler;
 
-		// 3. @Auth 받아오기
-		Auth auth = handlerMethod.getMethodAnnotation(Auth.class);
+		// 3. @Member 받아오기
+		Member auth = handlerMethod.getMethodAnnotation(Member.class);
 
-		// 4. 메소드에 @Auth 가 없는 경우 > 권한 인증이 필요없는 요청
+		// 4. 메소드에 @Member 가 없는 경우 > 권한 인증이 필요없는 요청
 		if (auth == null) return true;
 
-		// 5. @Auth 가 있는 경우이므로 세션이 있는지 체크 
+		// 5. @Member 가 있는 경우이므로 세션이 있는지 체크 
 		HttpSession session = request.getSession();
 		if(session == null) {
 			// 세션이 없을 경우 > Login 화면으로 이동 
-			response.sendRedirect(request.getContextPath() + "/member/login");
+			response.sendRedirect(request.getContextPath() + "/login");
 			return false;
 		}
 		
 		// 6. 세션이 존재하다면 유효한 유저인지 확인
 		MemberDto member = (MemberDto) session.getAttribute("userinfo");
 		if(member == null) {
-			response.sendRedirect(request.getContextPath() + "/member/login");
+			response.sendRedirect(request.getContextPath() + "/login");
 			return false;
 		}
