+++ ARORI/src/main/webapp/WEB-INF/views/member/login.jsp
@@ -34,8 +34,10 @@
 		<br><br>
 	</form>
 	<a href="find"><button>Find</button></a>
+	
 	<br><hr><br>
-<!-- 로그인 버튼 -->
+	
+	<!-- 로그인 버튼 -->
 	<div>
 		<div onclick="login(this);" id="google" class="firebaseui-idp-button mdl-button mdl-js-button mdl-button--raised firebaseui-idp-google firebaseui-id-idp-button">
 			<span class="firebaseui-idp-icon-wrapper">
